Re-enabled and added compiler flags. Fixed resulting warnings.
* Re-enabled "-Wconversion" (potential type conversion issues).
* Re-enabled "-Wswitch-enum" (ensure all cases explicitly handled).
* Added "-Wnon-virtual-dtor" (base classes with non-virtual destructors).
* Fixed all compiler warnings from enabling/adding those flags.
* Fixed -Wfloat-equal warnings with boost::math::epsilon_difference.

Other code simplifications and cleanup:

* Updated an error-prone std::equal call to safer C++14 version.
* Converted some manual floating-point formatting over to boost::format.

+++ datastore/importers/nmdb-import-pcap/PacketHelper.hpp
@@ -202,12 +202,12 @@ class PacketHelper {
     // UDP
     uint16_t getSrcPort(const UdpHeader*);
     uint16_t getDstPort(const UdpHeader*);
-    uint16_t getLength(const UdpHeader*);
+    size_t getLength(const UdpHeader*);
 
     // TCP
     uint16_t getSrcPort(const TcpHeader*);
     uint16_t getDstPort(const TcpHeader*);
-    uint8_t getLength(const TcpHeader*);
+    size_t getLength(const TcpHeader*);
 
     // DNS
     bool isResponse(const DnsHeader*);
