+++ playbook/nmdb-playbook/CommandRunnerSingleton.hpp
@@ -41,10 +41,10 @@ namespace netmeld::playbook {
     // =========================================================================
     private: // Variables will probably rarely appear at this scope
     protected: // Variables intended for internal/subclass API
-      bool                execute {false};
+      bool execute {false};
 
-      uint32_t            commandIdNumber {0};
-      std::set<uint32_t>  disabledCommands;
+      size_t commandIdNumber {0};
+      std::set<size_t> disabledCommands;
 
     public: // Variables should rarely appear at this scope
 
@@ -68,10 +68,10 @@ namespace netmeld::playbook {
     public: // Methods part of public API
       static CommandRunnerSingleton& getInstance();
 
-      void disableCommands(std::set<uint32_t> const&);
+      void disableCommands(std::set<size_t> const&);
       void setExecute(bool const);
 
-      bool isEnabled(uint32_t const) const;
+      bool isEnabled(size_t const) const;
 
       bool systemExec(std::string const&);
       void threadXtermExec(
